+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.6.13 (2020-03-25)
 
 - add keyboard support toggle as tray menu entry (#460)
+- add `CmdOrCtrl + F` as supported accelerator (#293, #454)
 - update disable keyboard support by defaut (#460)
 - update data to 3.10.0f
 - fix spacebar closes built in browser by disabling keyboard support (#461)